refactor: GA Custom Event 추가 & GA hook 분리
